+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [5.3.0] - 2024-05-31
 
 ### Added
 
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for threads of the WoW: Remix's `Cloak of Infinite Potential`.
     - Tracks amount of total threads (and amount gained per session).
     - Track remaining necessary threads for next "Infinite Power" level.
+
+### Changed
+- Optimized reputations under-the-hood, users should see no noticeable changes.
 
 ## [5.2.6] - 2024-05-07